+++ app/src/main/java/com/es/david/vacinas/activity/MainActivity.java
@@ -1,12 +1,8 @@
 package com.es.david.vacinas.activity;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.Snackbar;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.util.Log;
 import android.view.View;
 import android.support.design.widget.NavigationView;
 import android.support.v4.view.GravityCompat;
@@ -16,36 +12,19 @@ import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.widget.AdapterView;
-import android.widget.Toast;
 
 import com.es.david.vacinas.R;
-import com.es.david.vacinas.RecyclerItemClickListener;
-import com.es.david.vacinas.adapter.Adapter;
-import com.es.david.vacinas.modelo.Vacina;
-import com.google.firebase.database.DataSnapshot;
-import com.google.firebase.database.DatabaseError;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
-import com.google.firebase.database.ValueEventListener;
-
-import java.util.ArrayList;
-import java.util.List;
+import com.es.david.vacinas.fragmento.Campanhas;
+import com.es.david.vacinas.fragmento.Vacinas;
 
 public class MainActivity extends AppCompatActivity
         implements NavigationView.OnNavigationItemSelectedListener {
 
-
-    RecyclerView recyclerView;
-    private List<Vacina> listaVacinas = new ArrayList<>();
-    private DatabaseReference referencia = FirebaseDatabase.getInstance().getReference();
-    Adapter adapter;
-
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
+
+        setContentView(R.layout.activity_main2);
         Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
 
@@ -66,70 +45,6 @@ public class MainActivity extends AppCompatActivity
 
         NavigationView navigationView = (NavigationView) findViewById(R.id.nav_view);
         navigationView.setNavigationItemSelectedListener(this);
-
-        recyclerView = findViewById(R.id.recyclerV);
-
-        //Click evento
-
-        recyclerView.addOnItemTouchListener(
-                new RecyclerItemClickListener(
-                        getApplicationContext(),
-                        recyclerView,
-                        new RecyclerItemClickListener.OnItemClickListener() {
-                            @Override
-                            public void onItemClick(View view, int position) {
-                                Toast.makeText(
-                                        getApplicationContext(),
-                                        "HELLO WORLD",
-                                        Toast.LENGTH_SHORT
-                                ).show();
-                                Intent intent = new Intent(getApplicationContext(), DetalhesActivity.class);
-                                intent.putExtra("vacina", listaVacinas.get(position));
-                                startActivity(intent);
-                            }
-
-                            @Override
-                            public void onLongItemClick(View view, int position) {
-
-                            }
-
-                            @Override
-                            public void onItemClick(AdapterView<?> adapterView, View view, int i, long l) {
-
-                            }
-                        }
-                )
-
-        );
-
-
-
-        //// Fiberase
-
-        DatabaseReference vacinas = referencia.child("vacinas");
-
-        vacinas.addValueEventListener(new ValueEventListener() {
-            @Override
-            public void onDataChange(DataSnapshot dataSnapshot) {
-                for(DataSnapshot vacinaSnapshot : dataSnapshot.getChildren()) {
-                    Vacina vacina = vacinaSnapshot.getValue(Vacina.class);
-                    listaVacinas.add(vacina);
-                }
-//                adapter = new Adapter(listaVacinas);
-
-                RecyclerView.LayoutManager layoutManager = new LinearLayoutManager(getApplicationContext());
-                recyclerView.setLayoutManager(layoutManager);
-                recyclerView.setHasFixedSize(true);
-                recyclerView.setAdapter(adapter);
-            }
-
-            @Override
-            public void onCancelled(DatabaseError databaseError) {
-
-            }
-        });
-
-
     }
 
     @Override
@@ -145,7 +60,7 @@ public class MainActivity extends AppCompatActivity
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.main, menu);
+        getMenuInflater().inflate(R.menu.main2, menu);
         return true;
     }
 
@@ -170,11 +85,11 @@ public class MainActivity extends AppCompatActivity
         // Handle navigation view item clicks here.
         int id = item.getItemId();
 
-        if (id == R.id.nav_camera) {
+        if (id == R.id.campanhas) {
+            MainActivity.this.getSupportFragmentManager().beginTransaction().replace(R.id.containerID, Campanhas.newInstance()).commit();
             // Handle the camera action
-        } else if (id == R.id.nav_gallery) {
-
-        } else if (id == R.id.nav_logout) {
+        } else if (id == R.id.vacinas) {
+            MainActivity.this.getSupportFragmentManager().beginTransaction().replace(R.id.containerID, Vacinas.newInstance()).commit();
 
         }
 
@@ -182,6 +97,5 @@ public class MainActivity extends AppCompatActivity
         drawer.closeDrawer(GravityCompat.START);
         return true;
     }
-
-
 }
+
